pyproject.toml
[project]
name = "mpi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "langchain-ollama>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "linkup>=0.1.4",
    "mem0ai>=1.0.1",
    "pypdf>=6.4.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.38.0",
]
